apps/api/requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.34.3
sqlmodel==0.0.24
psycopg[binary]==3.2.9
pydantic==2.12.5
pydantic-settings==2.12.0
PyJWT[crypto]==2.10.1
bcrypt==4.3.0
email-validator==2.3.0
redis==6.4.0
python-multipart==0.0.20
